requirements.txt
flask==3.0.3
groq==0.24.0
python-dotenv==1.0.1
gunicorn==22.0.0
httpx==0.27.2